requirements.txt
flask
python-dotenv
openai
markdown
Pillow
piexif